requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
watchdog>=3.0.0
schedule>=1.2.0
chromadb>=0.4.0